requirements.txt

fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
cryptography
simpletransformers
transformers